requirements.txt
openai
gradio
pyttsx3
pydantic
python-dotenv